+++ R/LTTB_aggregator.R
@@ -31,14 +31,13 @@ LTTB_aggregator <- R6::R6Class(
     #' assuming 2 is the unit length of \code{x}
     #' (and 1 is always the unit length of \code{y}).
     #' The unit length is employed to calculate the area of the triangles.
-    #' @param interleave_gaps,coef_gap,NA_position,accepted_datatype,...
+    #' @param interleave_gaps,coef_gap,NA_position,...
     #' Arguments pass to the constructor of \code{aggregator} object.
     #' Note that \code{accepted_datatype} has default value.
     initialize = function(
       ...,
       nt_y_ratio = 1e9, x_y_ratio = 1.0,
-      interleave_gaps, coef_gap, NA_position,
-      accepted_datatype = c("numeric", "integer", "character", "factor", "logical")
+      interleave_gaps, coef_gap, NA_position
     ) {
       args <- c(as.list(environment()), list(...))
       do.call(super$initialize, args)
@@ -49,6 +48,7 @@ LTTB_aggregator <- R6::R6Class(
   ),
 
   private = list(
+    accepted_datatype = c("numeric", "integer", "character", "factor", "logical"),
     nt_y_ratio = 1,
     x_y_ratio = 1,
     aggregate_exec = function(x, y, n_out) {