+++ R/custom_stat_aggregator.R
@@ -30,12 +30,12 @@ custom_stat_aggregator <- R6::R6Class(
     #' By default, \code{TRUE}.
     #' @description
     #' Constructor of the Aggregator.
-    #' @param interleave_gaps,coef_gap,NA_position,accepted_datatype,...
+    #' @param interleave_gaps,coef_gap,NA_position,...
     #' Arguments pass to the constructor of \code{aggregator} object.
     initialize = function(
       ...,
       y_func = mean, x_mean = TRUE,
-      interleave_gaps, coef_gap, NA_position, accepted_datatype
+      interleave_gaps, coef_gap, NA_position
       ) {
       args <- c(as.list(environment()), list(...))
       do.call(super$initialize, args)
@@ -45,6 +45,7 @@ custom_stat_aggregator <- R6::R6Class(
 
   ),
   private = list(
+    accepted_datatype = c("numeric", "integer", "character", "factor", "logical"),
     y_func = NULL,
     x_mean = NULL,
     aggregate_exec = function(x, y, n_out) {