+++ R/nth_pnt_aggregator.R
@@ -16,17 +16,18 @@ nth_pnt_aggregator <- R6::R6Class(
   public = list(
     #' @description
     #' Constructor of the Aggregator.
-    #' @param interleave_gaps,coef_gap,NA_position,accepted_datatype,...
+    #' @param interleave_gaps,coef_gap,NA_position,...
     #' Arguments pass to the constructor of \code{aggregator} object.
     initialize = function(
       ...,
-      interleave_gaps, coef_gap, NA_position, accepted_datatype
+      interleave_gaps, coef_gap, NA_position
     ) {
       args <- c(as.list(environment()), list(...))
       do.call(super$initialize, args)
     }
   ),
   private = list(
+    accepted_datatype = c("numeric", "integer", "character", "factor", "logical"),
     aggregate_exec = function(x, y, n_out) {
       idx <- seq(1, length(x), max(1, ceiling(length(x) / n_out)))
       return(list(x = x[idx], y = y[idx]))