+++ R/rms_max_aggregator.R
@@ -1,4 +1,4 @@
-#' Aggregation which returns the ranges and nominal values
+#' Aggregation which returns the rms and maximum values
 #' within small data ranges
 #'
 #' @export
@@ -6,11 +6,11 @@
 #' @format An \code{R6::R6Class} object
 #' @description
 #' This aggregator divides the data into no-overlapping intervals
-#' and calculate specific statistics that represents the range and nominal
-#' values of the data, such as the max, min and mean.
+#' and calculate the root-mean-square and the maximum absolute
+#' values of the data, which may be helpful to understand the waveforms.
 #' @examples
 #' data(noise_fluct)
-#' agg <- range_stat_aggregator$new(
+#' agg <- rms_max_aggregator$new(
 #'   ylwr = min, y = mean, yupr = max, interleave_gaps = TRUE
 #' )
 #' d_agg <- agg$aggregate(nanotime::as.nanotime(noise_fluct$time), noise_fluct$f500, 100)
@@ -18,48 +18,29 @@
 #'   plotly::add_trace(x = d_agg$x, y = d_agg$ylwr, type = "scatter", mode = "lines")%>%
 #'   plotly::add_trace(x = d_agg$x, y = d_agg$yupr, type = "scatter", mode = "lines")
 #'
-range_stat_aggregator <- R6::R6Class(
-  "range_stat_aggregator",
+rms_max_aggregator <- R6::R6Class(
+  "rms_max_aggregator",
   inherit = rng_aggregator,
   public = list(
     #' @description
     #' Constructor of the aggregator.
     #' @param interleave_gaps,coef_gap,NA_position,...
     #' Arguments pass to the constructor of \code{aggregator} object.
-    #' @param yupr,y,ylwr Functions.
-    #' Statistical values are calculated using this function.
-    #' By default, \code{max, mean, min}, respectively.
-    #' Note that the NA values are omitted automatically.
+    #' @param use_abs
+    #' Boolean. If `TRUE`, the maximum absolute values are calculated.
     initialize = function(
       ...,
-      ylwr = min, y = mean, yupr = max,
-      interleave_gaps, coef_gap, NA_position
+      interleave_gaps, coef_gap, NA_position, use_abs = TRUE
     ) {
       args <- c(as.list(environment()), list(...))
       do.call(super$initialize, args)
 
-      assertthat::assert_that(
-        all(
-          c(
-            inherits(ylwr, "function"),
-            (inherits(y, "function") || is.null(y)),
-            inherits(yupr, "function")
-          )
-        ),
-        msg = "ylwr and yupr must be functions / y must be a function or NULL"
-      )
-      private$ylwr <- function(x) ylwr(na.omit(x))
-      private$yupr <- function(x) yupr(na.omit(x))
-      if (!is.null(y)) {
-        private$y    <- function(x) y(na.omit(x))
-      }
+      private$use_abs <- use_abs
     }
   ),
   private = list(
     accepted_datatype = c("numeric", "integer", "character", "factor", "logical"),
-    ylwr = NULL,
-    y    = NULL,
-    yupr = NULL,
+    use_abs = TRUE,
 
     aggregate_exec = function(x, y, n_out) {
 
@@ -72,17 +53,18 @@ range_stat_aggregator <- R6::R6Class(
         x_agg <- apply(x_mat, 2, mean, na.rm = TRUE)
       }
 
-      y_mat <- private$generate_matrix(y, n_out, remove_first_last = FALSE)
-
-      yupr_agg <- apply(y_mat, 2, private$yupr)
-      ylwr_agg <- apply(y_mat, 2, private$ylwr)
-      if (!is.null(private$y)) {
-        y_agg    <- apply(y_mat, 2, private$y)
+      if (private$use_abs) {
+        y_use <- abs(y)
       } else {
-        y_agg <- NA
+        y_use <- y
       }
+      y_mat <- private$generate_matrix(y_use, n_out, remove_first_last = FALSE)
 
-      return(list(x = x_agg, ylwr = ylwr_agg, y = y_agg, yupr = yupr_agg))
+      y_rms <- apply(y_mat, 2, function(x) sqrt(mean(x^2, na.rm=T)))
+      y_max <- apply(y_mat, 2, function(x) max(x, na.rm=T))
+
+
+      return(list(x = x_agg, ylwr = y_rms, y = y_rms, yupr = y_max))
     }
   )
 )